Polished up client
